Bug Fix: Deploying wars with unexpanded properties.
With newer Maven versions (2.2+) when a project is loaded from the repository getting the final build name gives the unexpanded version of ${project.artifactId}-${project.version} if it wasn't explicitly set in the pom.xml file.

The fix checks if that's happening and if so falls back to the deploy ID or the artefact ID if not.
I tried updating the version of maven this is using, but when attempting to switch to 2.2 it all blows up.